refactor: align public layout imports
