+++ resources/js/layouts/public/PublicLayout.tsx
