+++ resources/js/pages/agenda/Detail.tsx
@@ -1,6 +1,6 @@
 import { Head, Link, usePage } from '@inertiajs/react';
 import { ArrowLeft, CalendarCheck, CalendarClock, MapPin, Share2 } from 'lucide-react';
-import PublicLayout from '@/layouts/PublicLayout';
+import PublicLayout from '@/layouts/public/PublicLayout';
 import Breadcrumbs from '@/components/ui/Breadcrumbs';
 import type { EventSummary } from '@/features/content/types';
